+++ python_modules/libraries/dagster-celery/dagster_celery_tests/test_execute.py
@@ -97,9 +97,9 @@ def composite_pipeline():
 
 @solid(
     output_defs=[
-        OutputDefinition(Int, 'out_1', is_optional=True),
-        OutputDefinition(Int, 'out_2', is_optional=True),
-        OutputDefinition(Int, 'out_3', is_optional=True),
+        OutputDefinition(Int, 'out_1', is_required=False),
+        OutputDefinition(Int, 'out_2', is_required=False),
+        OutputDefinition(Int, 'out_3', is_required=False),
     ]
 )
 def foo(_):